1.4.0 1，注解添加： @EasyRedisCache @EasyRedisCacheClean 2，支持集合类型缓存快速使用 3，提供lua幂等操作接口

+++ src/main/java/cn/soilove/cache/annotations/EasyLock.java
@@ -27,7 +27,7 @@ public @interface EasyLock {
      * 超时时间 - 单位:秒
      * @return
      */
-    long timeout() default 60;
+    int timeout() default 60;
 
     /**
      * 是否自旋 - 默认非自旋
